Optimized event parsing code to do less fetching when parsing duplicate events.

+++ Nos/Service/EventProcessor.swift
@@ -15,7 +15,7 @@ enum EventProcessor {
         jsonObject: [String: Any],
         from relay: Relay?,
         in context: NSManagedObjectContext
-    ) throws -> Event {
+    ) throws -> Event? {
         let jsonData = try JSONSerialization.data(withJSONObject: jsonObject)
         let jsonEvent = try JSONDecoder().decode(JSONEvent.self, from: jsonData)
         return try parse(jsonEvent: jsonEvent, from: relay, in: context)
@@ -26,23 +26,36 @@ enum EventProcessor {
         from relay: Relay?,
         in parseContext: NSManagedObjectContext,
         skipVerification: Bool = false
-    ) throws -> Event {
-        let event = try Event.findOrCreate(jsonEvent: jsonEvent, relay: relay, context: parseContext)
+    ) throws -> Event? {
+        if let event = try Event.createIfNecessary(jsonEvent: jsonEvent, relay: relay, context: parseContext) {
         
-        guard let publicKey = event.author?.publicKey else {
-            throw EventError.missingAuthor
-        }
-        
-        if skipVerification == false {
-            guard try publicKey.verifySignature(on: event) else {
-                parseContext.delete(event)
-                Log.info("Invalid signature on event: \(jsonEvent)")
-                throw EventError.invalidSignature(event)
+            guard let publicKey = event.author?.publicKey else {
+                throw EventError.missingAuthor
             }
-            event.isVerified = true
+            
+            if skipVerification == false {
+                guard try publicKey.verifySignature(on: event) else {
+                    parseContext.delete(event)
+                    Log.info("Invalid signature on event: \(jsonEvent)")
+                    throw EventError.invalidSignature(event)
+                }
+                event.isVerified = true
+            }
+            
+            Log.debug("EventProcessor: parsed a new event")
+            return event
+            
+        // Verify that this event has been marked seen on the given relay.
+        } else if let relay, 
+            try parseContext.count(for: Event.event(by: jsonEvent.id, seenOn: relay)) == 0, 
+            let event = Event.find(by: jsonEvent.id, context: parseContext) {
+            event.markSeen(on: relay)
+            Log.debug("EventProcessor: marked an existing event seen")
+            return event
         }
         
-        return event
+        Log.debug("EventProcessor: skipping a duplicate event")
+        return nil
     }
     
     static func parse(jsonData: Data, from relay: Relay?, in context: NSManagedObjectContext) throws -> [Event] {
@@ -50,8 +63,9 @@ enum EventProcessor {
         var events = [Event]()
         for jsonEvent in jsonEvents {
             do {
-                let event = try parse(jsonEvent: jsonEvent, from: relay, in: context)
-                events.append(event)
+                if let event = try parse(jsonEvent: jsonEvent, from: relay, in: context) {
+                    events.append(event)
+                }
             } catch {
                 print("Error parsing eventJSON: \(jsonEvent): \(error.localizedDescription)")
             }
